+++ Bazzar/core-service/src/main/java/ru/bazzar/core/services/impl/OrderItemServiceImpl.java
@@ -1,4 +1,4 @@
-package ru.bazzar.core.servises.impl;
+package ru.bazzar.core.services.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.bazzar.core.entities.OrderItem;
 import ru.bazzar.core.repositories.OrderItemRepository;
-import ru.bazzar.core.servises.interf.OrderItemService;
+import ru.bazzar.core.services.interf.OrderItemService;
 
 @Service
 @Transactional
