+++ Bazzar/core-service/src/main/java/ru/bazzar/core/services/interf/SimpleService.java
@@ -1,4 +1,4 @@
-package ru.bazzar.core.servises.interf;
+package ru.bazzar.core.services.interf;
 //От него идет имплементация на остальные interf-сервисы(интерфейсы)
 public interface SimpleService<E,K> {
     E findById(K id);
